requirements.txt
opencv-contrib-python
pillow
mysql-connector-python
numpy
pandas
tkcalendar
